pyproject.toml
[tool.poetry]
name = "diet"
version = "0.1.0"
description = ""
authors = ["zdate <zdatezz0528@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.7"
pyyaml = "^6.0.1"
pydantic = "^2.7.2"
tabulate = "^0.9.0"

numpy = "^2.1.0"
scipy = "^1.14.1"
[tool.poetry.dev-dependencies]
black = "^24.8.0"  # Specify the version of Black you want to use

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
diet = "diet.main:cli"

[tool.black]
line-length = 79 # Default line length
skip-string-normalization = true  # Optional: skip normalization of string quotes
